Fix signing and verification using xmlsec and sign the assertion not the response
Using the xmlsec crate means we need to specify the path to the element being signed
and the ID attribute used to reference it.

The IdPs I've seen generally just sign the assertion with the option to sign the response. As sidre
progresses we can add a knob for signing the response.