Use xmlsec crate rather than generated bindings
This crate provides things such as allowing us to specify the xpath of what you
want to sign. This will allow us to sign the response and/or the assertion. Currently
it's using my fork of the crate because I've removed the panics and unwraps but once
I'm confident in my changes I'll try and get it merged and will switch it back to mainline.

It's also just a first cut of moving to the bindings. The interface kind of sucks because
you're having to pass all these xpaths and namespaces in and that should really be abstracted
away.